sockaddr family
In FreeBSD kernel, if the address family is unknown, it automatically
recognised as AF_INET6 because the address family of "dst" is
AF_INET6.  So, This is not critical but mask.sa_family should be set
in theory.

"route minitor" command cannot parse the routing message correctly at
least when I want to watch the routing socket.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>